ajout login, profile et chaque task/liste appartiennne a un user

+++ src/components/list/listes-controller.js
@@ -4,7 +4,7 @@ import tasksModel from "#components/tasks/tasks-model.js";
 
 export async function getAll(ctx) {
     try {
-        ctx.body = await listeModel.find()
+        ctx.body = await listeModel.find({createBy: ctx.state.user.id});
     } catch (e) {
         ctx.badRequest({message: e.message})
     }
@@ -13,6 +13,11 @@ export async function getAll(ctx) {
 export async function getOne(ctx) {
     try {
         let liste = await listeModel.findById(ctx.params.id).lean();
+        if(liste.createBy.toString() !== ctx.state.user.id) {
+            ctx.unauthorized();
+            return;
+        }
+
         liste.tasks = await tasksModel.findByListId(ctx.params.id);
 
         ctx.body = liste;
@@ -24,13 +29,15 @@ export async function getOne(ctx) {
 export async function create(ctx) {
     try {
         const listeValidationSchema = Joi.object({
-            title: Joi.string().required()
+            title: Joi.string().required(),
+            description: Joi.string().optional()
         });
 
         const { error } = listeValidationSchema.validate(ctx.request.body);
         if(error) throw new Error(error);
 
         let listeObj = ctx.request.body;
+        listeObj.createBy = ctx.state.user.id;
 
         await listeModel.create(listeObj);
         ctx.response.status = 201;
@@ -42,12 +49,19 @@ export async function create(ctx) {
 export async function edit(ctx) {
     try {
         const listeValidationSchema = Joi.object({
-            title: Joi.string(),
+            title: Joi.string().required(),
+            description: Joi.string().optional(),
         });
 
         const { error } = listeValidationSchema.validate(ctx.request.body);
         if(error) throw new Error(error);
 
+        const liste = await listeModel.find({_id: ctx.params.id});
+
+        if(!liste[0] || liste[0].createBy.toString() !== ctx.state.user.id) {
+            ctx.unauthorized();
+            return;
+        }
 
         let exempleObj = ctx.request.body;
 
@@ -60,6 +74,13 @@ export async function edit(ctx) {
 
 export async function remove(ctx) {
     try {
+        const liste = await listeModel.find({_id: ctx.params.id});
+
+        if(!liste[0] || liste[0].createBy.toString() !== ctx.state.user.id) {
+            ctx.unauthorized();
+            return;
+        }
+
         await listeModel.findOneAndRemove({_id: ctx.params.id});
         await tasksModel.deleteMany()
         ctx.response.status = 200;
