+++ src/components/tasks/tasks-controller.js
@@ -1,9 +1,10 @@
 import tasksModel from "#components/tasks/tasks-model.js";
+import listesModel from "#components/list/listes-model.js";
 import Joi from "joi";
 
 export async function getAll(ctx) {
     try {
-        ctx.body = await tasksModel.find()
+        ctx.body = await tasksModel.find({createBy: ctx.state.user.id})
     } catch (e) {
         ctx.badRequest({message: e.message})
     }
@@ -12,6 +13,12 @@ export async function getAll(ctx) {
 export async function getAllByListId(ctx) {
     try {
         if(!ctx.params.listId) throw new Error("No id supplied");
+        const list = await listesModel.find({_id: ctx.params.listId});
+        if(!list[0] || list[0].createBy.toString() !== ctx.state.user.id) {
+            ctx.unauthorized();
+            return;
+        }
+
         const tasks = await tasksModel.findByListId(ctx.params.listId);
         ctx.ok(tasks);
     } catch (e) {
@@ -21,7 +28,14 @@ export async function getAllByListId(ctx) {
 
 export async function getOne(ctx) {
     try {
-        ctx.body = await tasksModel.find({_id: ctx.params.id});
+        const task = await tasksModel.findById(ctx.params.id);
+
+        if(ctx.state.user.id !== task.createBy.toString()) {
+            ctx.unauthorized();
+            return;
+        }
+
+        ctx.ok(task);
     } catch (e) {
         ctx.badRequest({message: e.message})
     }
@@ -40,6 +54,7 @@ export async function create(ctx) {
         if(error) throw new Error(error);
 
         let tasksObj = ctx.request.body;
+        tasksObj.createBy = ctx.state.user.id;
 
         await tasksModel.create(tasksObj);
         ctx.response.status = 201;
@@ -60,6 +75,11 @@ export async function edit(ctx) {
         const { error } = tasksValidationSchema.validate(ctx.request.body);
         if(error) throw new Error(error);
 
+        let task = await tasksModel.find({_id: ctx.params.id});
+        if(task[0].createBy.toString() !== ctx.state.user.id) {
+            ctx.unauthorized();
+            return;
+        }
 
         let tasksObj = ctx.request.body;
 
@@ -72,6 +92,13 @@ export async function edit(ctx) {
 
 export async function remove(ctx) {
     try {
+        const task = await tasksModel.find({_id: ctx.params.id});
+
+        if(!task[0] || task[0].createBy.toString() !== ctx.state.user.id) {
+            ctx.unauthorized();
+            return;
+        }
+
         await tasksModel.findOneAndRemove({_id: ctx.params.id});
         ctx.response.status = 200;
     } catch (e) {
